itmo_bot/requirements.txt

python-telegram-bot==20.3
pandas
pdfplumber
python-dotenv
requests
beautifulsoup4
